requirements.txt
asgiref==3.2.7
backports.csv==1.0.7
certifi==2020.4.5.1
cloudinary==1.20.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.6
django-import-export==2.1.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
psycopg2==2.8.5
PyPDF2==1.26.0
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.42
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
tablib==2.0.0
urllib3==1.25.9
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0

